refactor(curate): inline exif resolvers into value factory
